requirements.txt
GitPython===3.1.31
setuptools-scm===7.1.0